feat: компонент render
