+++ src/helpers/render.js
@@ -1,0 +1,28 @@
+const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+};
+
+const InsertionPosition = {
+  AFTERBEGIN: `afterbegin`,
+  AFTEREND: `afterend`,
+  BEFOREEND: `beforeend`
+};
+
+const render = (container, element, position) => {
+  switch (position) {
+    case InsertionPosition.AFTERBEGIN:
+      container.prepend(element);
+      break;
+    case InsertionPosition.AFTEREND:
+      container.after(element);
+      break;
+    case InsertionPosition.BEFOREEND:
+      container.append(element);
+      break;
+  }
+};
+
+export {createElement, InsertionPosition, render};
