+++ src/helpers/utils.js
@@ -10,31 +10,4 @@ const getISOStringDate = (date) => {
   return isoDate.toISOString();
 };
 
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-const InsertionPosition = {
-  AFTERBEGIN: `afterbegin`,
-  AFTEREND: `afterend`,
-  BEFOREEND: `beforeend`
-};
-
-const render = (container, element, position) => {
-  switch (position) {
-    case InsertionPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case InsertionPosition.AFTEREND:
-      container.after(element);
-      break;
-    case InsertionPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-export {castTimeFormat, getFormatTime24H, getISOStringDate, createElement, InsertionPosition, render};
+export {castTimeFormat, getFormatTime24H, getISOStringDate};
